Move GeoTiffRasterSources into a separate geotiff package
