+++ vlm/src/main/resources/reference.conf
@@ -1,7 +1,7 @@
 gdal.cache {
   maximumSize: 1000
   enableDefaultRemovalListener: true
-  valuesType: weak
+  valuesType: Weak
   enabled: true
   withShutdownHook: true
 }
