+++ vlm/src/test/scala/geotrellis/contrib/vlm/LayoutTileSourceSpec.scala
@@ -16,16 +16,12 @@
 
 package geotrellis.contrib.vlm
 
-import geotrellis.contrib.vlm.gdal._
+import geotrellis.contrib.vlm.geotiff._
 import geotrellis.raster._
-import geotrellis.raster.reproject.Reproject
 import geotrellis.raster.io.geotiff.reader._
 
-import geotrellis.proj4._
 import geotrellis.spark._
-import geotrellis.spark.io.hadoop._
 import geotrellis.spark.tiling._
-import geotrellis.spark.testkit._
 import geotrellis.raster.testkit.RasterMatchers
 
 import org.scalatest._
