+++ vlm/src/test/scala/geotrellis/contrib/vlm/RasterRegionSpec.scala
@@ -1,7 +1,7 @@
 package geotrellis.contrib.vlm
 
+import geotrellis.contrib.vlm.geotiff._
 import geotrellis.raster.io.geotiff._
-
 import geotrellis.proj4._
 import geotrellis.spark._
 import geotrellis.spark.tiling._
