+++ vlm/src/test/scala/geotrellis/contrib/vlm/geotiff/GeoTiffReprojectRasterSourceSpec.scala
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
  
-package geotrellis.contrib.vlm
+package geotrellis.contrib.vlm.geotiff
 
+import geotrellis.contrib.vlm._
 import geotrellis.raster._
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
 import geotrellis.raster.resample._
